Filled components + views
All components have their values, besides WalletForm which should pop after user signs up or for MyWalletView inside ProfilePage (we need to know which values it should hold).
All views are ready to start connecting to the api, there is still need to polish html order in most components but it is ready for a starting point.